Remove array with empty object as only field
Not filling in anything into the processes form results in a weird data
structure - an array with an empty object as oly item. This array now
gets removed.

I feel we should remove all empty fields like empty strings and arrays
as well. Feels kind of inconsistent as it is now.